% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_BLR_2_sampler.R
\name{hmc_base_sampler_BLR_2}
\alias{hmc_base_sampler_BLR_2}
\title{HMC sampler for base level for logistic regression model}
\usage{
hmc_base_sampler_BLR_2(
  nsamples,
  warmup,
  data_split,
  C,
  prior_means,
  prior_variances,
  seed = sample.int(.Machine$integer.max, 1),
  n_cores = parallel::detectCores(),
  output = F
)
}
\arguments{
\item{nsamples}{number of samples per node}

\item{warmup}{number of burn in iterations}

\item{data_split}{list of length C where each item is a list where for c=1,...,C,
data_split[\link{c}]$y is a vector of responses and data_split[\link{c}]$X
is the design matrix}

\item{C}{number of sub-posteriors}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{seed}{seed number for random number generation}

\item{n_cores}{number of cores to use}

\item{output}{boolean value: defaults to T, determines whether or not
to print output to console}
}
\value{
samples from the sub-posterior targets for the split data sets
for the logistic regression model
}
\description{
Sample for base level for logistic regression
}
