% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_BLR_sampler.R
\name{hmc_sample_BLR}
\alias{hmc_sample_BLR}
\title{HMC sampler for Logistic Regression model}
\usage{
hmc_sample_BLR(
  full_data_count,
  C,
  prior_means,
  prior_variances,
  power = 1,
  iterations,
  warmup,
  chains,
  seed = sample.int(.Machine$integer.max, 1),
  output = F
)
}
\arguments{
\item{full_data_count}{a matrix or dataframe of the unique data with their
corresponding counts}

\item{C}{number of sub-posterior}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{power}{the exponent of the posterior (default is 1)}

\item{iterations}{number of iterations per chain}

\item{warmup}{number of burn in iterations}

\item{chains}{number of chains}

\item{seed}{seed number for random number generation}

\item{output}{boolean value: defaults to T, determines whether or not to
print output to console}
}
\value{
samples from the (sub-)posterior target for the logistic regression model
}
\description{
Sample from (sub-)posterior using Stan
}
