% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_BLR_2_sampler.R
\name{hmc_sample_BLR_2}
\alias{hmc_sample_BLR_2}
\title{HMC sampler for Logistic Regression model}
\usage{
hmc_sample_BLR_2(
  y,
  X,
  C,
  prior_means,
  prior_variances,
  iterations,
  warmup,
  chains,
  seed = sample.int(.Machine$integer.max, 1),
  output = F
)
}
\arguments{
\item{y}{vector of responses}

\item{X}{design matrix}

\item{C}{number of sub-posterior}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{iterations}{number of iterations per chain}

\item{warmup}{number of burn in iterations}

\item{chains}{number of chains}

\item{seed}{seed number for random number generation}

\item{output}{boolean value: defaults to T, determines whether or not to
print output to console}
}
\value{
samples from the (sub-)posterior target for the logistic regression model
}
\description{
Sample from (sub-)posterior using Stan
}
